doen ticket 7.3
At each iteration, the loop prompts the visitor to enter a number greater than 100 (but a maximum of 10 iterations of the loop). If the visitor enters a number less than one hundred, it prompts them to enter it again, and so on. If the user enters more than a hundred, the text or loop finishes all iterations, the function displays the last user input to the console and terminates the function. And split each action into a separate function