done ticket 12.1
There are two buttons on the page. When clicking on the first button, the user has to enter a link in the prompt, and when clicking on the second button, the user is redirected to another site (using the previously entered link).
I added a validation check for the lead link and a style for the buttons.